# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Companion API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TravelerTravelerType =
  (typeof TravelerTravelerType)[keyof typeof TravelerTravelerType];

export const TravelerTravelerType = {
  adult: "adult",
  child: "child",
} as const;

export interface Traveler {
  id: number;
  userId?: string | null;
  name: string;
  birthDate?: string | null;
  travelerType: TravelerTravelerType;
  relationship?: string | null;
  foodPreferences: string[];
  activityPreferences: string[];
  accessibilityNeeds?: string | null;
  notes?: string | null;
  ageYears?: number | null;
  ageMonths?: number | null;
  ageDisplay?: string | null;
}

export type CreateTravelerBodyTravelerType =
  (typeof CreateTravelerBodyTravelerType)[keyof typeof CreateTravelerBodyTravelerType];

export const CreateTravelerBodyTravelerType = {
  adult: "adult",
  child: "child",
} as const;

export interface CreateTravelerBody {
  userId?: string | null;
  name: string;
  birthDate?: string | null;
  travelerType: CreateTravelerBodyTravelerType;
  relationship?: string | null;
  foodPreferences?: string[] | null;
  activityPreferences?: string[] | null;
  accessibilityNeeds?: string | null;
  notes?: string | null;
}

export interface UpdateTravelerParams {
  id: number;
}

export interface TravelPreferences {
  id: number;
  seatPreference?: string | null;
  mealPreference?: string | null;
  frequentFlyerNumbers?: string | null;
  passportNotes?: string | null;
  accessibilityNeeds?: string | null;
  hotelPreferences?: string | null;
  travelInsuranceNotes?: string | null;
  additionalNotes?: string | null;
  travelStyles?: string[] | null;
  travelStyleTags?: string[] | null;
  luxuryIndexMin?: number | null;
  luxuryIndexMax?: number | null;
  priceValueWeight?: number | null;
  notes?: string | null;
}

export interface UpsertPreferencesBody {
  seatPreference?: string | null;
  mealPreference?: string | null;
  frequentFlyerNumbers?: string | null;
  passportNotes?: string | null;
  accessibilityNeeds?: string | null;
  hotelPreferences?: string | null;
  travelInsuranceNotes?: string | null;
  additionalNotes?: string | null;
  travelStyles?: string[] | null;
  travelStyleTags?: string[] | null;
  luxuryIndexMin?: number | null;
  luxuryIndexMax?: number | null;
  priceValueWeight?: number | null;
  notes?: string | null;
}

export interface TravelSummary {
  children: Child[];
  preferences?: TravelPreferences;
  totalTravelers: number;
  hasPreferences: boolean;
}

export type ScoreReviewsBodySource =
  (typeof ScoreReviewsBodySource)[keyof typeof ScoreReviewsBodySource];

export const ScoreReviewsBodySource = {
  tripadvisor: "tripadvisor",
  google: "google",
} as const;

export interface ScoreReviewsBody {
  propertyId: string;
  source: ScoreReviewsBodySource;
  /** @minItems 1 */
  reviews: string[];
}

export type ReviewScoreRawClaudeResponse = { [key: string]: unknown } | null;

export interface ReviewScore {
  propertyId: string;
  source: string;
  reviewHash: string;
  reviewText: string;
  tags?: string[] | null;
  luxuryValueScore?: number | null;
  foodieScore?: number | null;
  ecoScore?: number | null;
  adventurousMenuScore?: number | null;
  sentiment?: string | null;
  oneLineSummary?: string | null;
  rawClaudeResponse?: ReviewScoreRawClaudeResponse;
  cachedAt: string;
}

export interface MatchResult {
  matchScore: number;
  topReviews: ReviewScore[];
  tagSummary: string[];
}

export interface TravelerSummaryItem {
  name: string;
  travelerType: string;
  relationship?: string | null;
  ageYears?: number | null;
  ageMonths?: number | null;
  foodPreferences: string[];
  activityPreferences: string[];
}

export interface FamilySummary {
  adults: TravelerSummaryItem[];
  children: TravelerSummaryItem[];
  travelerCount: number;
  partyDescription: string;
}

export interface PreferencesSummary {
  travelStyleTags: string[];
  luxuryIndexMin: number;
  luxuryIndexMax: number;
  priceValueWeight: number;
  notes?: string | null;
}

export interface AutoFillPayload {
  adults: number;
  children: number;
  childAges: number[];
  partyDescription: string;
}

export interface WeightVector {
  luxuryValue: number;
  foodie: number;
  eco: number;
  adventurousMenu: number;
}

export interface ReviewProfile {
  weightVector: WeightVector;
  description: string;
}

export interface ExtendedTravelSummary {
  family: FamilySummary;
  preferences: PreferencesSummary;
  autoFillPayload: AutoFillPayload;
  reviewProfile: ReviewProfile;
}

export type MatchReviewsParams = {
  property_id: string;
  user_id?: string;
};