# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Travel Companion API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: travelers
    description: Travel party management
  - name: preferences
    description: Travel preferences management
  - name: reviews
    description: Review scoring and matching
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /travelers:
    get:
      operationId: listTravelers
      tags: [travelers]
      summary: List all travelers in the party
      responses:
        "200":
          description: List of travelers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Traveler"
    post:
      operationId: createTraveler
      tags: [travelers]
      summary: Add a traveler
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTravelerBody"
      responses:
        "201":
          description: Created traveler
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Traveler"

  /travelers/{id}:
    put:
      operationId: updateTraveler
      tags: [travelers]
      summary: Update a traveler
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTravelerBody"
      responses:
        "200":
          description: Updated traveler
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Traveler"
    delete:
      operationId: deleteTraveler
      tags: [travelers]
      summary: Delete a traveler
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /preferences:
    get:
      operationId: getPreferences
      tags: [preferences]
      summary: Get travel preferences
      responses:
        "200":
          description: Travel preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelPreferences"
    put:
      operationId: upsertPreferences
      tags: [preferences]
      summary: Save travel preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertPreferencesBody"
      responses:
        "200":
          description: Saved preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelPreferences"

  /summary:
    get:
      operationId: getTravelSummary
      tags: [children]
      summary: Get full travel summary including ages at today
      responses:
        "200":
          description: Travel summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TravelSummary"

  /summary/{userId}:
    get:
      operationId: getExtendedSummary
      tags: [children]
      summary: Rich summary for browser extension — includes auto-fill payload and review profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Extended travel summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtendedTravelSummary"

  /reviews/score:
    post:
      operationId: scoreReviews
      tags: [reviews]
      summary: Score one or more property reviews using AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScoreReviewsBody"
      responses:
        "200":
          description: Array of scored reviews (from cache or freshly computed)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewScore"
        "429":
          description: Rate limit exceeded

  /reviews/match:
    get:
      operationId: matchReviews
      tags: [reviews]
      summary: Compute a weighted match score for a property against user preferences
      parameters:
        - name: property_id
          in: query
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Match result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResult"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Traveler:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
          nullable: true
        name:
          type: string
        birthDate:
          type: string
          format: date
          nullable: true
        travelerType:
          type: string
          enum: [adult, child]
        relationship:
          type: string
          nullable: true
        foodPreferences:
          type: array
          items:
            type: string
        activityPreferences:
          type: array
          items:
            type: string
        accessibilityNeeds:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        ageYears:
          type: integer
          nullable: true
        ageMonths:
          type: integer
          nullable: true
        ageDisplay:
          type: string
          nullable: true
      required:
        - id
        - name
        - travelerType
        - foodPreferences
        - activityPreferences

    CreateTravelerBody:
      type: object
      properties:
        userId:
          type: string
          nullable: true
        name:
          type: string
        birthDate:
          type: string
          format: date
          nullable: true
        travelerType:
          type: string
          enum: [adult, child]
        relationship:
          type: string
          nullable: true
        foodPreferences:
          type: array
          items:
            type: string
          nullable: true
        activityPreferences:
          type: array
          items:
            type: string
          nullable: true
        accessibilityNeeds:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - name
        - travelerType

    UpdateTravelerParams:
      type: object
      properties:
        id:
          type: integer
      required:
        - id

    TravelPreferences:
      type: object
      properties:
        id:
          type: integer
        seatPreference:
          type: string
          nullable: true
        mealPreference:
          type: string
          nullable: true
        frequentFlyerNumbers:
          type: string
          nullable: true
        passportNotes:
          type: string
          nullable: true
        accessibilityNeeds:
          type: string
          nullable: true
        hotelPreferences:
          type: string
          nullable: true
        travelInsuranceNotes:
          type: string
          nullable: true
        additionalNotes:
          type: string
          nullable: true
        travelStyles:
          type: array
          items:
            type: string
          nullable: true
        travelStyleTags:
          type: array
          items:
            type: string
          nullable: true
        luxuryIndexMin:
          type: integer
          nullable: true
        luxuryIndexMax:
          type: integer
          nullable: true
        priceValueWeight:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - id

    UpsertPreferencesBody:
      type: object
      properties:
        seatPreference:
          type: string
          nullable: true
        mealPreference:
          type: string
          nullable: true
        frequentFlyerNumbers:
          type: string
          nullable: true
        passportNotes:
          type: string
          nullable: true
        accessibilityNeeds:
          type: string
          nullable: true
        hotelPreferences:
          type: string
          nullable: true
        travelInsuranceNotes:
          type: string
          nullable: true
        additionalNotes:
          type: string
          nullable: true
        travelStyles:
          type: array
          items:
            type: string
          nullable: true
        travelStyleTags:
          type: array
          items:
            type: string
          nullable: true
        luxuryIndexMin:
          type: integer
          nullable: true
        luxuryIndexMax:
          type: integer
          nullable: true
        priceValueWeight:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true

    TravelSummary:
      type: object
      properties:
        children:
          type: array
          items:
            $ref: "#/components/schemas/Child"
        preferences:
          $ref: "#/components/schemas/TravelPreferences"
        totalTravelers:
          type: integer
        hasPreferences:
          type: boolean
      required:
        - children
        - totalTravelers
        - hasPreferences

    ScoreReviewsBody:
      type: object
      properties:
        propertyId:
          type: string
        source:
          type: string
          enum: [tripadvisor, google]
        reviews:
          type: array
          items:
            type: string
          minItems: 1
      required:
        - propertyId
        - source
        - reviews

    ReviewScore:
      type: object
      properties:
        propertyId:
          type: string
        source:
          type: string
        reviewHash:
          type: string
        reviewText:
          type: string
        tags:
          type: array
          items:
            type: string
          nullable: true
        luxuryValueScore:
          type: integer
          nullable: true
        foodieScore:
          type: integer
          nullable: true
        ecoScore:
          type: integer
          nullable: true
        adventurousMenuScore:
          type: integer
          nullable: true
        sentiment:
          type: string
          nullable: true
        oneLineSummary:
          type: string
          nullable: true
        rawClaudeResponse:
          type: object
          nullable: true
        cachedAt:
          type: string
          format: date-time
      required:
        - propertyId
        - source
        - reviewHash
        - reviewText
        - cachedAt

    MatchResult:
      type: object
      properties:
        matchScore:
          type: number
        topReviews:
          type: array
          items:
            $ref: "#/components/schemas/ReviewScore"
        tagSummary:
          type: array
          items:
            type: string
      required:
        - matchScore
        - topReviews
        - tagSummary

    TravelerSummaryItem:
      type: object
      properties:
        name:
          type: string
        travelerType:
          type: string
        relationship:
          type: string
          nullable: true
        ageYears:
          type: integer
          nullable: true
        ageMonths:
          type: integer
          nullable: true
        foodPreferences:
          type: array
          items:
            type: string
        activityPreferences:
          type: array
          items:
            type: string
      required:
        - name
        - travelerType
        - foodPreferences
        - activityPreferences

    FamilySummary:
      type: object
      properties:
        adults:
          type: array
          items:
            $ref: "#/components/schemas/TravelerSummaryItem"
        children:
          type: array
          items:
            $ref: "#/components/schemas/TravelerSummaryItem"
        travelerCount:
          type: integer
        partyDescription:
          type: string
      required:
        - adults
        - children
        - travelerCount
        - partyDescription

    PreferencesSummary:
      type: object
      properties:
        travelStyleTags:
          type: array
          items:
            type: string
        luxuryIndexMin:
          type: integer
        luxuryIndexMax:
          type: integer
        priceValueWeight:
          type: integer
        notes:
          type: string
          nullable: true
      required:
        - travelStyleTags
        - luxuryIndexMin
        - luxuryIndexMax
        - priceValueWeight

    AutoFillPayload:
      type: object
      properties:
        adults:
          type: integer
        children:
          type: integer
        childAges:
          type: array
          items:
            type: integer
        partyDescription:
          type: string
      required:
        - adults
        - children
        - childAges
        - partyDescription

    WeightVector:
      type: object
      properties:
        luxuryValue:
          type: number
        foodie:
          type: number
        eco:
          type: number
        adventurousMenu:
          type: number
      required:
        - luxuryValue
        - foodie
        - eco
        - adventurousMenu

    ReviewProfile:
      type: object
      properties:
        weightVector:
          $ref: "#/components/schemas/WeightVector"
        description:
          type: string
      required:
        - weightVector
        - description

    ExtendedTravelSummary:
      type: object
      properties:
        family:
          $ref: "#/components/schemas/FamilySummary"
        preferences:
          $ref: "#/components/schemas/PreferencesSummary"
        autoFillPayload:
          $ref: "#/components/schemas/AutoFillPayload"
        reviewProfile:
          $ref: "#/components/schemas/ReviewProfile"
      required:
        - family
        - preferences
        - autoFillPayload
        - reviewProfile