# lib/api-zod/src/generated/types/createTravelerBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTravelerBodyTravelerType } from "./createTravelerBodyTravelerType";

export interface CreateTravelerBody {
  userId?: string | null;
  name: string;
  birthDate?: Date | null;
  travelerType: CreateTravelerBodyTravelerType;
  relationship?: string | null;
  foodPreferences?: string[] | null;
  activityPreferences?: string[] | null;
  accessibilityNeeds?: string | null;
  notes?: string | null;
}