# lib/api-zod/src/generated/types/familySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { TravelerSummaryItem } from "./travelerSummaryItem";

export interface FamilySummary {
  adults: TravelerSummaryItem[];
  children: TravelerSummaryItem[];
  travelerCount: number;
  partyDescription: string;
}