# lib/api-zod/src/generated/types/traveler.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { TravelerTravelerType } from "./travelerTravelerType";

export interface Traveler {
  id: number;
  userId?: string | null;
  name: string;
  birthDate?: Date | null;
  travelerType: TravelerTravelerType;
  relationship?: string | null;
  foodPreferences: string[];
  activityPreferences: string[];
  accessibilityNeeds?: string | null;
  notes?: string | null;
  ageYears?: number | null;
  ageMonths?: number | null;
  ageDisplay?: string | null;
}