# lib/api-zod/src/generated/types/travelerSummaryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Travel Companion API
 * OpenAPI spec version: 0.1.0
 */

export interface TravelerSummaryItem {
  name: string;
  travelerType: string;
  relationship?: string | null;
  ageYears?: number | null;
  ageMonths?: number | null;
  foodPreferences: string[];
  activityPreferences: string[];
}